dpfj tnwjddk

+++ korchart.py
@@ -136,7 +136,7 @@ def run(ticker, com_name):
     #######################################################
     rim_price = int(value_df.iloc[4].replace(',','').replace('원', ''))
     current_price = int(value_df.iloc[3].replace(',','').replace('원', ''))
-    a_yield = int(value_df.iloc[7].replace('%',''))
+    a_yield = float(value_df.iloc[7].replace('%',''))
     col1, col2, col3 = st.columns(3)
     col1.metric(label="Price", value = current_price, delta=rim_price)
     col2.metric(label="PBR", value =round(float(value_df.iloc[13]),2), delta=round(float(value_df.iloc[14]),2))
